UI Update 1.0
Whenever user edits the textView or textField, the dateLabel gets updated to the current live date.  It then saves the current date in Realm.  Unless the user edits the textView or textField, dateLabel displays the updated current date that is stored in Realm whenever the user exits/restarts the app.